feat(main.models): Add AbstractSkill
AbstractSkill is an abstract model for representing skills. It includes a field for when the skill was first learned, and properties for calculating and displaying years of experience. It also includes a property that returns the static path to the skill's icon. The model inherits name and slug fields from AbstractNamedModel, and validation logic from ValidatedModelMixin.