feat(website.models): Add CacheAllManager and AbstractCacheAllModel
The CacheAll abstract model and manager are a pair of a model and manager that are intended to add caching to a child model on its all() queryset. The cache will automatically be cleared when an item is saved or when an item is deleted. The cache is initially set/retrieved in the manager's all() method.